+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-## Next Release
+## Next release
+- Cache last n blocks by using a new Besu flag --cache-last-blocks=n
 
 ### Breaking Changes